Refactor logging mechanism to use PSR-3 compatible logger and remove deprecated debug mode
