+++ composer.json
@@ -11,7 +11,8 @@
         "php": "^8.1",
         "ext-mbstring": "*",
         "ext-curl": "*",
-        "psr/simple-cache": "^3.0"
+        "psr/simple-cache": "^3.0",
+        "psr/log": "^3.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^9"
