+++ src/Bolt.php
@@ -23,6 +23,10 @@ final class Bolt
     private array $protocolVersions = [];
     private int $packStreamVersion = 1;
 
+    /**
+     * Debug mode - print sent and received buffers as HEX
+     * @deprecated Use PSR-3 compatible logger instead
+     */
     public static bool $debug = false;
 
     public function __construct(private IConnection $connection)
@@ -204,9 +208,6 @@ final class Bolt
      */
     private function handshake(): string
     {
-        if (self::$debug)
-            echo 'HANDSHAKE';
-
         $this->connection->write(chr(0x60) . chr(0x60) . chr(0xb0) . chr(0x17) . $this->packProtocolVersions());
 
         $bytes = $this->connection->read(4);
